Project overtaking!
sinefuse has abandoned both this GH account and the t-sp.in project for health reasons. As such, I, ameliaRTA, now have full ownership over the repository (and the GH account for that matter! He's okay with it). I figured it'd be a good idea to change a few small details in the files.

Not much else will change, anything that exists right now will still exist.